refactor(kolme): extract live provider module ownership (#1952)
